--- a/Code/GraphMol/Bond.h
+++ b/Code/GraphMol/Bond.h
@@ -17,6 +17,7 @@
 // FIX: grn...
 #include <Query/QueryObjects.h>
 #include <RDGeneral/types.h>
+#include <RDGeneral/RDProps.h>
 #include <GraphMol/details.h>
 #include <boost/foreach.hpp>
 
@@ -43,7 +44,7 @@ typedef boost::shared_ptr<Atom> ATOM_SPTR;
           clients who need to store extra data on Bond objects.
 
 */
-class Bond {
+class Bond : public RDProps{
   friend class RWMol;
   friend class ROMol;
 
@@ -288,141 +289,6 @@ class Bond {
     return *dp_stereoAtoms;
   };
 
-  // ------------------------------------
-  //  Local Property Dict functionality
-  //  FIX: at some point this stuff should go in a mixin class
-  // ------------------------------------
-  //! returns a list with the names of our \c properties
-  STR_VECT getPropList() const { return dp_props->keys(); }
-
-  //! sets a \c property value
-  /*!
-     \param key the name under which the \c property should be stored.
-         If a \c property is already stored under this name, it will be
-         replaced.
-     \param val the value to be stored
-     \param computed (optional) allows the \c property to be flagged
-         \c computed.
-   */
-  template <typename T>
-  void setProp(const char *key, T val, bool computed = false) const {
-    // if(!dp_props) dp_props = new Dict();
-    std::string what(key);
-    setProp(what, val, computed);
-  }
-  //! \overload
-  template <typename T>
-  void setProp(const std::string &key, T val, bool computed = false) const {
-    // setProp(key.c_str(),val);
-    if (computed) {
-      STR_VECT compLst;
-      getPropIfPresent(detail::computedPropName, compLst);
-      if (std::find(compLst.begin(), compLst.end(), key) == compLst.end()) {
-        compLst.push_back(key);
-        dp_props->setVal(detail::computedPropName, compLst);
-      }
-    }
-    dp_props->setVal(key, val);
-  }
-
-  //! allows retrieval of a particular property value
-  /*!
-
-     \param key the name under which the \c property should be stored.
-         If a \c property is already stored under this name, it will be
-         replaced.
-     \param res a reference to the storage location for the value.
-
-     <b>Notes:</b>
-       - if no \c property with name \c key exists, a KeyErrorException will be
-     thrown.
-       - the \c boost::lexical_cast machinery is used to attempt type
-     conversions.
-         If this fails, a \c boost::bad_lexical_cast exception will be thrown.
-
-  */
-  template <typename T>
-  void getProp(const char *key, T &res) const {
-    PRECONDITION(dp_props, "getProp called on empty property dict");
-    dp_props->getVal(key, res);
-  }
-  //! \overload
-  template <typename T>
-  void getProp(const std::string &key, T &res) const {
-    PRECONDITION(dp_props, "getProp called on empty property dict");
-    dp_props->getVal(key, res);
-  }
-
-  //! \Overload
-  template <typename T>
-  T getProp(const char *key) const {
-    return dp_props->getVal<T>(key);
-  }
-  //! \overload
-  template <typename T>
-  T getProp(const std::string &key) const {
-    return dp_props->getVal<T>(key);
-  }
-
-  //! returns whether or not we have a \c property with name \c key
-  //!  and assigns the value if we do
-
-  template <typename T>
-  bool getPropIfPresent(const char *key, T &res) const {
-    return dp_props->getValIfPresent(key, res);
-  }
-  //! \overload
-  template <typename T>
-  bool getPropIfPresent(const std::string &key, T &res) const {
-    return dp_props->getValIfPresent(key, res);
-  }
-
-  //! returns whether or not we have a \c property with name \c key
-  bool hasProp(const char *key) const {
-    if (!dp_props) return false;
-    return dp_props->hasVal(key);
-  };
-  //! \overload
-  bool hasProp(const std::string &key) const {
-    if (!dp_props) return false;
-    return dp_props->hasVal(key);
-  };
-
-  //! clears the value of a \c property
-  /*!
-     <b>Notes:</b>
-       - if no \c property with name \c key exists, a KeyErrorException
-         will be thrown.
-       - if the \c property is marked as \c computed, it will also be removed
-         from our list of \c computedProperties
-  */
-  void clearProp(const char *key) const {
-    std::string what(key);
-    clearProp(what);
-  };
-  //! \overload
-  void clearProp(const std::string &key) const {
-    STR_VECT compLst;
-    if (getPropIfPresent(detail::computedPropName, compLst)) {
-      STR_VECT_I svi = std::find(compLst.begin(), compLst.end(), key);
-      if (svi != compLst.end()) {
-        compLst.erase(svi);
-        dp_props->setVal(detail::computedPropName, compLst);
-      }
-    }
-    dp_props->clearVal(key);
-  }
-
-  //! clears all of our \c computed \c properties
-  void clearComputedProps() const {
-    STR_VECT compLst;
-    if (getPropIfPresent(detail::computedPropName, compLst)) {
-      BOOST_FOREACH (const std::string &sv, compLst) { dp_props->clearVal(sv); }
-      compLst.clear();
-      dp_props->setVal(detail::computedPropName, compLst);
-    }
-  }
-
   //! calculates any of our lazy \c properties
   /*!
     <b>Notes:</b>
@@ -443,7 +309,6 @@ class Bond {
   atomindex_t d_index;
   atomindex_t d_beginAtomIdx, d_endAtomIdx;
   ROMol *dp_mol;
-  Dict *dp_props;
   INT_VECT *dp_stereoAtoms;
 
   void initBond();