--- a/Code/GraphMol/RWMol.h
+++ b/Code/GraphMol/RWMol.h
@@ -214,11 +214,10 @@ class RWMol : public ROMol {
     d_bondBookmarks.clear();
     d_graph.clear();
     d_confs.clear();
-    if (dp_props) {
-      dp_props->reset();
-      STR_VECT computed;
-      dp_props->setVal(detail::computedPropName, computed);
-    }
+    dp_props.reset();
+    STR_VECT computed;
+    dp_props.setVal(detail::computedPropName, computed);
+
     if (dp_ringInfo) dp_ringInfo->reset();
   };