--- a/Code/GraphMol/Wrap/Bond.cpp
+++ b/Code/GraphMol/Wrap/Bond.cpp
@@ -293,7 +293,10 @@ struct bond_wrapper {
         .def("GetPropNames", &Bond::getPropList, (python::arg("self")),
              "Returns a list of the properties set on the Bond.\n\n")
 
-        .def("GetPropsAsDict", GetPropsAsDict<Bond>, (python::arg("self")),
+        .def("GetPropsAsDict", GetPropsAsDict<Bond>, (python::arg("self"),
+                                                      python::arg("includePrivate") = true,
+                                                      python::arg("includeComputed") = true
+                                                      ),
              "Returns a dictionary of the properties set on the Bond.\n"
              " n.b. some properties cannot be converted to python types.\n")