… Code/GraphMol/ROMol.cpp
… Code/GraphMol/ROMol.h
… Code/GraphMol/Wrap/Mol.cpp
… Code/GraphMol/Wrap/props.hpp
… Code/GraphMol/Wrap/rough_test.py
… Code/JavaWrappers/Atom.i
… Code/JavaWrappers/RDProps.i
… Code/JavaWrappers/ROMol.i
… Code/JavaWrappers/gmwrapper/GraphMolJava.i
… Code/RDGeneral/CMakeLists.txt
… Code/RDGeneral/Dict.cpp
… Code/RDGeneral/Dict.h
… Code/RDGeneral/Invariant.h
… Code/RDGeneral/RDAny.h
… Code/RDGeneral/RDProps.h
… Code/RDGeneral/RDValue-doublemagic.h
… Code/RDGeneral/RDValue-taggedunion.h
… Code/RDGeneral/RDValue.h
… Code/RDGeneral/testDict.cpp
… Code/RDGeneral/testRDValue.cpp
… Code/RDGeneral/types.cpp
… Code/RDGeneral/types.h