This is an API compliant version of the current rdany system, but uses a lot less memory in practice.

There are two versions of RDAny for evaluation, one is a simple tagged union (64 + 16 bits in size) and the other is a NaN tagged value (64 bits).

The NaN tagged value is neat, and surprisingly works well, it, however, exploits both the NaN bit structure in IEEE floating point AND the fact that current OSes don't use more that 48 bits to access memory.  This technique is used in some Javascript engines, but still makes me feel dirty inside.

The default is the tagged union, and this may indeed win out in the end especially if we use integer keys instead of strings (this is another PR for  later).

Both systems fall back to boost::any storage, so no functionality is lost.  Instead of using boost::any_cast, a new cast is added rdany_cast, however end users should not have to use this.

Additionally, a new class has been created RDProps, which, when subclassed, enabled property support for any class.  Atoms/Bonds and Molecules now use RDProps cleaning up a lot of redundant code.

Finally the types.h namespace has been documented and cleaned up a bit.