@greglandrum 2016-05-09
Just to have the reference material somewhere...

I wanted to compare this to the current state of master in terms of memory usage and performance.
Here's runtime of the standard RDKit timing script (`$RDBASE/Regress/Scripts/timings.py`):
```
master:
times:  0.4 || 0.2 || 0.0 || 0.8 || 0.8 || 0.0 || 2.4 || 2.5 || 0.1 || 0.1 || 4.8 || 0.2 || 4.8 || 13.9 || 1.1 || 0.7 || 27.3

this branch:
times:  0.4 || 0.2 || 0.0 || 0.7 || 0.8 || 0.0 || 2.4 || 2.5 || 0.1 || 0.1 || 4.4 || 0.2 || 4.7 || 13.8 || 1.0 || 0.7 || 26.4

```

Since there's not enough room to see much in those numbers, I did another test where I created 10K molecules from SMILES (using a set of ZINC molecules) without freeing the memory.

Timing:
```
master:

this branch:
master:
      1.897781s wall, 1.840000s user + 0.050000s system = 1.890000s CPU (99.6%)

this branch:
      1.748275s wall, 1.700000s user + 0.040000s system = 1.740000s CPU (99.5%)
```

Memory usage for the 10K molecules:
```
master:
  ==19114== HEAP SUMMARY:
  ==19114==     in use at exit: 176,510,434 bytes in 3,697,421 blocks
  ==19114==   total heap usage: 17,992,392 allocs, 14,294,971 frees, 842,503,033 bytes allocated
  ==19114== 
  ==19114== LEAK SUMMARY:
  ==19114==    definitely lost: 780,408 bytes in 14,231 blocks
  ==19114==    indirectly lost: 175,416,543 bytes in 3,679,056 blocks
  ==19114==      possibly lost: 40,239 bytes in 843 blocks
  ==19114==    still reachable: 273,244 bytes in 3,291 blocks
  ==19114==         suppressed: 0 bytes in 0 blocks

this branch:
  ==11621== HEAP SUMMARY:
  ==11621==     in use at exit: 151,587,998 bytes in 2,923,633 blocks
  ==11621==   total heap usage: 15,449,221 allocs, 12,525,588 frees, 780,042,466 bytes allocated
  ==11621== 
  ==11621== LEAK SUMMARY:
  ==11621==    definitely lost: 3,604,440 bytes in 139,301 blocks
  ==11621==    indirectly lost: 147,510,675 bytes in 2,777,391 blocks
  ==11621==      possibly lost: 199,175 bytes in 3,648 blocks
  ==11621==    still reachable: 273,708 bytes in 3,293 blocks
  ==11621==         suppressed: 0 bytes in 0 blocks
```