@greglandrum 2016-05-09
It's all really lost (there's no free'ing anywhere). I wouldn't worry about it.
Still, the code for the test is here: https://gist.github.com/greglandrum/8aeea5e108ca43008c8a9a9eea871c0e