@greglandrum 2016-05-09
Ok, I've read through all of this now and tested on windows and linux (including 32bit Centos5) and it looks good.
I think it's ok to merge once the couple of small suggestions I made are in (particularly the code duplication thing).