@bp-kelley 2016-05-09
Give me a few days to test on PowerPC Debian.  The emulator is really slow :)

----
Brian Kelley

> On May 9, 2016, at 8:51 AM, Greg Landrum <notifications@github.com> wrote:
> 
> Ok, I've read through all of this now and tested on windows and linux (including 32bit Centos5) and it looks good.
> I think it's ok to merge once the couple of small suggestions I made are in (particularly the code duplication thing).
> 
> —
> You are receiving this because you authored the thread.
> Reply to this email directly or view it on GitHub
>