@greglandrum 2016-05-09
It's probably faster to try it on a raspberry PI