@bp-kelley 2016-05-09
You just want rdkit on a raspberry :)