@greglandrum 2016-05-09
Na, @janholstjensen already did that. Presented it at the first RDKit UGM too. :-)