@bp-kelley 2016-05-17
@greglandrum the perfectionist in me is trying out big endian architectures.  I finally got qemu running, if not I have a 2004 era PowerPC iMac floating around :)