@bp-kelley 2016-05-25
@greglandrum should be ok now.  Once the is merged we can try to get the property serialization to work for pod data.